healthiar: Quantifying and Monetizing Health Impacts Attributable to Exposure

This R package has been developed with a focus on air pollution and noise but can be applied to other exposures. The initial development has been funded by the European Union project BEST-COST. Disclaimer: It is work in progress and the developers are not liable for any calculation errors or inaccuracies resulting from the use of this package. Selection of relevant references (in chronological order): WHO (2003) https://www.who.int/publications/i/item/9241546204, Murray et al. (2003) doi:10.1186/1478-7954-1-1 , Miller & Hurley (2003) doi:10.1136/jech.57.3.200 , Steenland & Armstrong (2006) doi:10.1097/01.ede.0000229155.05644.43 , WHO (2011) https://iris.who.int/items/723ab97c-5c33-4e3b-8df1-744aa5bc1c27, GBD 2019 Risk Factors Collaborators (2020) doi:10.1016/S0140-6736(20)30752-2 .

See also

Author

Maintainer: Alberto Castro alberto.castrofernandez@swisstph.ch (ORCID)

Authors:

Axel Luyten (ORCID)

Other contributors:

Arno Pauwels (ORCID) [contributor]

Liliana Vazquez Fernandez (ORCID) [contributor]

Gianni Ardielli (ORCID) [contributor]

Iracy Pimenta (ORCID) [contributor]

Susanne Breitner (ORCID) [contributor]

Carl Baravelli (ORCID) [contributor]

Vanessa Gorasso (ORCID) [contributor]

Maria Lepnurm (ORCID) [contributor]

Andreia Novais (ORCID) [contributor]

Maria Jose Rueda Lopez (ORCID) [contributor]

Minh Nhat Pham (ORCID) [contributor]

Ana Barbosa (ORCID) [contributor]

Joao Vasco Santos (ORCID) [contributor]

Anette Kocbach Bolling (ORCID) [contributor]