Aggregate health impacts from multiple exposures

This function aggregates health impacts from multiple exposures to environmental stressors.

multiexpose(
  output_attribute_exp_1,
  output_attribute_exp_2,
  exp_name_1,
  exp_name_2,
  approach_multiexposure = "additive"
)

Arguments

output_attribute_exp_1, output_attribute_exp_2

Output of attribute() for exposure 1 and 2, respectively. Baseline health data and population must be identical in outputs 1 and 2.

exp_name_1, exp_name_2

String referring to the name of the environmental exposures 1 and 2

approach_multiexposure

String specifying the multiple exposures approach to be used in the assessment. Options: "additive" (default), "multiplicative" or "combined".

Value

This function returns a list containing:

1) health_main (tibble) containing the main results;

impact (numeric column) attributable health burden/impact

pop_fraction (numeric column) population attributable fraction; only applicable in relative risk assessments

And many more

2) health_detailed (list) containing detailed (and interim) results.

results_raw (tibble) containing results for each combination of input uncertainty

results_by_geo_id_micro (tibble) containing results for each geographic unit under analysis (specified in geo_id_micro argument)

input_table (tibble) containing the inputs to each relevant argument

input_args (list) containing all the argument inputs used in the background

Details

Sources

For more information on the additive and combined approaches see Steenland & Armstrong 2006 (https://doi.org/10.1097/01.ede.0000229155.05644.43).

For more information on the multiplicative approach see Jerrett et al. 2013 (https://doi.org/10.1164/rccm.201303-0609OC).

Author

Alberto Castro & Axel Luyten

Examples

# Goal: determine aggregated health impacts from multiple exposures
# Step 1: create assessment with exposure 1
output_attribute_exp_1 <- attribute_health(
  erf_shape = "log_linear",
  rr_central = 1.369,
  rr_increment = 10,
  exp_central = 8.85,
  cutoff_central = 5,
  bhd_central = 30747
)
output_attribute_exp_1$health_main$impact
#> [1] 3501.962
# Step 2: create assessment with exposure 2
output_attribute_exp_2 <- attribute_mod(
  output_attribute = output_attribute_exp_1,
  exp_central = 10.9,
  rr_central = 1.031
)
output_attribute_exp_2$health_main$impact
#> [1] 548.8641
# Step 3: aggregate impacts of the two assessments
results <- multiexpose(
  output_attribute_exp_1 = output_attribute_exp_1,
  output_attribute_exp_2 = output_attribute_exp_2,
  exp_name_1 = "pm2.5",
  exp_name_2 = "no2",
  approach_multiexposure = "multiplicative"
)
results$health_main$impact
#> [1] 3988.312